Implement a pair of GitHub actions to measure output size and report in the PR.
"Measure" action is supposed to be used in a "pull_request" workflow. It
checks out a reference app and builds it with GopherJS from the PR and
captures output size for non-minified, minified and gzipped build modes.

The same set of measurements is performed against GopherJS from the
target branch and, if the target branch is not master, against master.
The corresponding output artifact sizes are then compared and rendered
into a report in Markdown format.

"Publish" action then obtains the rendered report and posts it as a
comment in the PR. If there are previous reports already in the PR, they
will be hidden to reduce distraction.